+++ tests/conftest.py
@@ -7,6 +7,7 @@ from sqlalchemy.orm import Session
 
 from src.db.models.document import DocumentDB
 from src.db.models.index_version import IndexVersionDB
+from src.db.models.run import IngestionRunDB
 
 load_dotenv()
 
@@ -33,4 +34,5 @@ def database_session(database_engine):
 
         session.execute(delete(DocumentDB))
         session.execute(delete(IndexVersionDB))
+        session.execute(delete(IngestionRunDB))
         session.commit()